fix(ros): pass extra-input-globs
